Cleaning project and little refactor
